Initial preparation for "next generation" camerad 2.0, adds source for Reed's FITS engine to repo but doesn't utilize it yet, cleans up and renames a few things but no functional changes. So far, this will compile and continue to function as the "main" branch does.

+++ utils/sendcmd.cpp
@@ -32,7 +32,7 @@ int main(int argc, char *argv[]) {
   if ( argc==1 ) { std::cout << usage(); return 0; }
 
   try {
-  for ( size_t i=1; i<argc; ++i ) {
+  for ( int i=1; i<argc; ++i ) {
     if (argv[i][0] == '-') {
       switch ( argv[i][1] ) {
         case 'h' : if ( i+1 < argc ) hostname = argv[++i];
